repair selling land
returned some predicates to their old state, added building to OurLand because some multiPolygons aren't saved as land but owned by us as a building.
NOTE: AT this moment the bot does sell the housing once but does not try to sell it again when declined, the bot also sells roads which is not really what we planned for, should be fixed soon.